requirements.txt

wget~=3.2
tqdm~=4.64.1
wandb~=0.13.10
numpy~=1.24.2
scipy~=1.10.1
torch~=1.13.1
PyQt5~=5.15.9
PyYAML~=6.0
pandas
matplotlib~=3.7.0
torchvision~=0.14.1
opencv-python-headless

opencv-python~=4.7.0.72
Pillow~=9.4.0
vispy~=0.13.0